HubUrl now available on default and payload only processed if subscription is available
Now it is possible to not specify the hubUrl in the default.ini. In this case a default
url is used with the following scheme: '<OntoWikiUrl>pubsub/hubbub/'.
When the payload is delivered to the callback-action, the pubsub-extension checks if
a subscription for the given topic is saved in the configuration. Only in this case
the onExternalFeedDidChange-event is called.